Proof of concept session storage.
This uses the index and service-before pages to prove out automatically
storing form data into a session, where it will be available in any
templates further down the flow.

There is a set of defaults that can be specified, so the flow should
still work and present useful data even if the user skips an input
field.

The session store used at the moment clears on restarts (ie on changes),
which may not be ideal.